+++ R/tglkmeans-package.R
@@ -2,18 +2,20 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#'
 #' @import dplyr
 #' @import ggplot2
-#' @importFrom purrr set_names
-#' @importFrom stats cor dist hclust var
-#' @importFrom utils capture.output
+#' @importFrom cli cli_abort cli_warn cli_alert cli_alert_info cli_alert_success
 #' @importFrom grDevices colorRampPalette dev.off
+#' @importFrom purrr set_names
+#' @importFrom Rcpp sourceCpp
+#' @importFrom RcppParallel RcppParallelLibs
+#' @importFrom stats cor dist hclust var
 #' @importFrom stats cutree
 #' @importFrom tgstat tgs_cor
 #' @importFrom tgstat tgs_dist
 #' @importFrom tibble remove_rownames column_to_rownames rownames_to_column has_rownames rowid_to_column as_tibble tibble
-#'
+#' @importFrom utils capture.output
 #' @useDynLib tglkmeans
-#' @importFrom Rcpp sourceCpp
 ## usethis namespace: end
 NULL