+++ src/RcppExports.cpp
@@ -34,8 +34,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // TGL_kmeans_cpp
-List TGL_kmeans_cpp(const StringVector& ids, DataFrame& mat, const int& k, const String& metric, const double& max_iter, const double& min_delta, const bool& random_seed, const int& seed);
-RcppExport SEXP _tglkmeans_TGL_kmeans_cpp(SEXP idsSEXP, SEXP matSEXP, SEXP kSEXP, SEXP metricSEXP, SEXP max_iterSEXP, SEXP min_deltaSEXP, SEXP random_seedSEXP, SEXP seedSEXP) {
+List TGL_kmeans_cpp(const StringVector& ids, DataFrame& mat, const int& k, const String& metric, const double& max_iter, const double& min_delta, const bool& use_cpp_random, const int& seed);
+RcppExport SEXP _tglkmeans_TGL_kmeans_cpp(SEXP idsSEXP, SEXP matSEXP, SEXP kSEXP, SEXP metricSEXP, SEXP max_iterSEXP, SEXP min_deltaSEXP, SEXP use_cpp_randomSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -45,9 +45,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const String& >::type metric(metricSEXP);
     Rcpp::traits::input_parameter< const double& >::type max_iter(max_iterSEXP);
     Rcpp::traits::input_parameter< const double& >::type min_delta(min_deltaSEXP);
-    Rcpp::traits::input_parameter< const bool& >::type random_seed(random_seedSEXP);
+    Rcpp::traits::input_parameter< const bool& >::type use_cpp_random(use_cpp_randomSEXP);
     Rcpp::traits::input_parameter< const int& >::type seed(seedSEXP);
-    rcpp_result_gen = Rcpp::wrap(TGL_kmeans_cpp(ids, mat, k, metric, max_iter, min_delta, random_seed, seed));
+    rcpp_result_gen = Rcpp::wrap(TGL_kmeans_cpp(ids, mat, k, metric, max_iter, min_delta, use_cpp_random, seed));
     return rcpp_result_gen;
 END_RCPP
 }