+++ tests/testthat/test-clustering.R
@@ -22,15 +22,15 @@ clustering_ok <- function(data, res, nclust, ndims, order = TRUE) {
     expect_equal(nrow(data), sum(res$size$n))
 }
 
-context("Missing data")
+# Missing data:
 test_that("Stop when there are rows which contain only missing data", {
     data <- as.data.frame(simulate_data(n = 100, sd = 0.3, nclust = 30, frac_na = NULL))
     data[3, -1] <- NA
     data[4, -1] <- NA
-    expect_error(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = FALSE, seed = 60427), "The following rows contain only missing values: 3,4")
+    expect_error(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427))
 })
 
-context("Matrix input")
+# Matrix input:
 test_that("Do not fail when input is matrix", {
     nclust <- 30
     ndims <- 5
@@ -43,7 +43,7 @@ test_that("Do not fail when input is matrix", {
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 })
 
-context("Rownames")
+# Rownames:
 test_that("Use rownames if exists", {
     nclust <- 30
     ndims <- 5
@@ -52,13 +52,10 @@ test_that("Use rownames if exists", {
         as.data.frame() %>%
         select(id, starts_with("V")) %>%
         mutate(id = paste0("id_", id)) %>%
-        column_to_rownames("id")
+        tibble::column_to_rownames("id")
 
     res <- TGL_kmeans_tidy(data, 30, id_column = FALSE, metric = "euclid", verbose = FALSE, seed = 60427)
     clustering_ok(data, res, nclust, ndims, order = FALSE)
-
-    expect_warning(res1 <- TGL_kmeans_tidy(data, 30, id_column = TRUE, metric = "euclid", verbose = FALSE, seed = 60427))
-    clustering_ok(data, res1, nclust, ndims, order = FALSE)
 })
 
 test_that("Dot not fail when rownames do not exist", {
@@ -68,7 +65,7 @@ test_that("Dot not fail when rownames do not exist", {
     data <- data %>%
         select(starts_with("V")) %>%
         as.data.frame()
-    data <- remove_rownames(data)
+    data <- tibble::remove_rownames(data)
 
     res <- TGL_kmeans_tidy(data, 30, id_column = FALSE, metric = "euclid", verbose = FALSE, seed = 60427)
     clustering_ok(data, res, nclust, ndims, order = FALSE)
@@ -76,7 +73,7 @@ test_that("Dot not fail when rownames do not exist", {
     res_non_tidy <- TGL_kmeans(data, 30, id_column = FALSE, metric = "euclid", verbose = FALSE, seed = 60427)
 })
 
-context("Metrics")
+# Metrics:
 test_that("Pearson metric works", {
     nclust <- 30
     ndims <- 5
@@ -93,13 +90,13 @@ test_that("Spearman metric works", {
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 })
 
-context("Correct output")
+# Correct output:
 
 test_that("all ids and clusters are present", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE)
 
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 })
@@ -108,8 +105,8 @@ test_that("non tidy version works", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, seed = 60427)
-    res_tidy <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, seed = 60427)
+    res <- TGL_kmeans(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427)
+    res_tidy <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427)
 
     clustering_ok(data, res_tidy, nclust, ndims, order = FALSE)
 
@@ -126,9 +123,9 @@ test_that("hclust intra cluster works", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, hclust_intra_clusters = TRUE, parallel = FALSE, seed = 60427)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, hclust_intra_clusters = TRUE, parallel = FALSE, seed = 60427)
     clustering_ok(data, res, nclust, ndims, order = TRUE)
-    res_non_tidy <- TGL_kmeans(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, hclust_intra_clusters = TRUE, parallel = FALSE, , seed = 60427)
+    res_non_tidy <- TGL_kmeans(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, hclust_intra_clusters = TRUE, parallel = FALSE, , seed = 60427)
 
     expect_equal(res_non_tidy$order, res$order$order)
 })
@@ -137,7 +134,7 @@ test_that("add_to_data works", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, add_to_data = TRUE)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, add_to_data = TRUE)
 
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 
@@ -148,17 +145,17 @@ test_that("add_to_data works", {
         as.data.frame() %>%
         select(id, starts_with("V")) %>%
         mutate(id = paste0("id_", id)) %>%
-        column_to_rownames("id")
+        tibble::column_to_rownames("id")
 
     res <- TGL_kmeans_tidy(data, 30, id_column = FALSE, metric = "euclid", verbose = FALSE, seed = 60427, add_to_data = TRUE)
-    expect_equal(res$data %>% select(starts_with("V")), data %>% select(starts_with("V")))
+    expect_equal(res$data %>% select(starts_with("V")), data %>% select(starts_with("V")), ignore_attr = TRUE)
 })
 
 test_that("reorder func works when set to mean", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, reorder_func = mean)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, reorder_func = mean)
 
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 })
@@ -167,43 +164,43 @@ test_that("reorder func works when set to NULL", {
     nclust <- 30
     ndims <- 5
     data <- simulate_data(n = 200, sd = 0.3, dims = 5, nclust = nclust, frac_na = 0.05)
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, reorder_func = NULL)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, reorder_func = NULL)
 
     clustering_ok(data, res, nclust, ndims, order = FALSE)
 })
 
-context("Verbosity")
+# Verbosity:
 test_that("quiet if verbose is turned off", {
     data <- simulate_data(n = 100, sd = 0.3, nclust = 30, frac_na = NULL)
-    expect_silent(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = FALSE, seed = 60427))
+    expect_silent(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427))
 })
 
 test_that("not quiet when verbose is turned on", {
     data <- simulate_data(n = 100, sd = 0.3, nclust = 30, frac_na = NULL)
-    expect_message(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = TRUE, seed = 60427))
+    expect_output(TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = TRUE, seed = 60427))
 })
 
 test_that("Log is saved when 'keep_log' is turned on", {
     data <- simulate_data(n = 100, sd = 0.3, nclust = 30, frac_na = NULL)
-    expect_warning(res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = TRUE, seed = 60427, keep_log = TRUE))
-    expect_warning(res <- TGL_kmeans(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = TRUE, seed = 60427, keep_log = TRUE))
+    expect_warning(res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, id_column = TRUE, metric = "euclid", verbose = TRUE, seed = 60427, keep_log = TRUE))
+    expect_warning(expect_warning(res <- TGL_kmeans(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = TRUE, seed = 60427, keep_log = TRUE)))
 
-    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = FALSE, seed = 60427, keep_log = TRUE)
+    res <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427, keep_log = TRUE)
     expect_type(res$log, "character")
-    res <- TGL_kmeans(data %>% select(id, starts_with("V")), 30, metric = "euclid", verbose = FALSE, seed = 60427, keep_log = TRUE)
+    res <- TGL_kmeans(data %>% select(id, starts_with("V")), 30, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427, keep_log = TRUE)
     expect_type(res$log, "character")
 })
 
-context("Random seed")
+# Random seed:
 test_that("setting the seed returns reproducible results", {
     nclust <- 30
     data <- simulate_data(n = 100, sd = 0.3, nclust = nclust, frac_na = NULL)
-    res1 <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, seed = 60427)
-    res2 <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", verbose = FALSE, seed = 60427)
+    res1 <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427)
+    res2 <- TGL_kmeans_tidy(data %>% select(id, starts_with("V")), nclust, metric = "euclid", id_column = TRUE, verbose = FALSE, seed = 60427)
     expect_true(all(res1$centers[, -1] == res2$centers[, -1]))
 })
 
-context("Correct Classification (low dim)")
+# Correct Classification (low dim):
 test_that("clustering is reasonable (low dim): euclid", {
     test_params <- expand.grid(n = c(100), sd = c(0.05, 0.1, 0.3), nclust = c(5, 30, 100), dims = c(2, 10)) %>% filter(nclust < n)
     apply(test_params, 1, function(x) {
@@ -218,7 +215,7 @@ test_that("clustering with NA is reasonable (low dim): euclid", {
     })
 })
 
-context("Correct Classification (high dim)")
+# Correct Classification (high dim):
 test_that("clustering is reasonable (high dim): euclid", {
     skip_on_cran()
     test_params <- expand.grid(n = c(500), sd = c(0.3), nclust = c(5, 30), dims = c(300)) %>% filter(nclust < n)
@@ -235,7 +232,7 @@ test_that("clustering with NA is reasonable (high dim): euclid", {
     })
 })
 
-context("Data simulation")
+# Data simulation:
 test_that("true_clust column is not added when add_true_clust is FALSE", {
     data <- simulate_data(n = 100, sd = 0.3, nclust = 30, frac_na = NULL, add_true_clust = FALSE)
     expect_true(!("true_clust" %in% colnames(data)))