+++ NEWS.md
@@ -1,16 +1,22 @@
+# tglkmeans 0.4.0
+
+* Default of `id_column` parameter was changed to `FALSE`. Note that this is a breaking change, and if you want to use an id column, you need to set it explicitly to `TRUE`.
+* Use R random number generator instead of C++11 random number generator. For backwards compatibility, the old random number generator can be used by setting `use_cpp_random` to `TRUE`.
+* Added parallelization using `RcppParallel`. 
+
 # tglkmeans 0.3.12
 
 * Added validity checks for `k` and the number of observations. 
 
-# tgkmeans 0.3.11
+# tglkmeans 0.3.11
 
 * Changed pkgdoc, see: https://github.com/r-lib/roxygen2/issues/1491.
 
-# tgkmeans 0.3.10
+# tglkmeans 0.3.10
 
 * Removed broken link to one of the references in the description.
 
-# tgkmeans 0.3.9
+# tglkmeans 0.3.9
 
 * Remove empty clusters. This may happen when the number of clusters is larger than the number of observations, and currently caused an error in the reordering step.